requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
bs4==0.0.1
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.2
channels==2.4.0
channels-redis==3.0.1
chardet==4.0.0
charset-normalizer==2.0.6
cloudinary==1.26.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
daphne==2.5.0
dateparser==1.0.0
discord.py==1.7.3
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.8
django-cloudinary-storage==0.3.0
django-cors-headers==3.8.0
django-dotenv==1.4.2
django-filter==21.1
django-heroku==0.3.1
django-jazzmin==2.4.8
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
filelock==3.0.12
firebase-admin==5.2.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
gunicorn==20.1.0
hiredis==1.1.0
html5lib==1.1
httplib2==0.20.2
hyperlink==20.0.1
idna==2.10
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jsonschema==4.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
msgpack==1.0.2
multidict==5.1.0
openapi-codec==1.3.2
packaging==21.0
Pillow==8.3.2
proto-plus==1.19.8
protobuf==3.19.1
psycopg==3.0b1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2021.8.28
requests==2.26.0
rsa==4.8
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
simplejson==3.17.6
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.2
static3==0.7.0
swagger-spec-validator==2.7.3
twilio==7.4.0
Twisted==21.7.0

txaio==20.4.1
typing-extensions==3.10.0.2
tzdata==2021.1
tzlocal==3.0
uritemplate==4.1.1
urllib3==1.26.6
virtualenv==20.4.6
webencodings==0.5.1
whitenoise==5.3.0
xmltodict==0.12.0
yarl==1.6.3
zope.interface==5.1.0